Extract cross-stage overlap attribution from iterate_rpc_inference
The ~400-line [CROSS_STAGE_OVERLAP_SUMMARY] analysis lived inline in the
micro-batch merge branch of iterate_rpc_inference, deep inside an already
2300-line generator. It is pure reporting over collected timestamps, so it
moves to overlap_attribution.summarize_cross_stage_overlap(); the call
site passes the step's overlap records and accumulator. block_functions
shrinks by ~390 lines. No behavior change.
